+++ quill/test/PatternFormatterTest.cpp
@@ -178,6 +178,72 @@ TEST(PatternFormatter, custom_pattern_timestamp_precision_none)
   EXPECT_EQ(formatted_string, expected_string);
 }
 
+TEST(PatternFormatter, custom_pattern_timestamp_strftime_reallocation_on_format_string_2)
+{
+  // set a timestamp_format that will cause timestamp _formatted_date to re-allocate.
+  PatternFormatter custom_pattern_formatter{
+    QUILL_STRING(
+      "%(ascii_time) [%(thread)] %(filename):%(lineno) LOG_%(level_name) %(logger_name) - "
+      "%(message) [%(function_name)]"),
+    "%FT%T.%Qus%z%Z", Timezone::GmtTime};
+
+  std::chrono::nanoseconds ts{1579815761099220000};
+  char const* thread_id = "31341";
+  std::string const logger_name = "test_logger";
+  StaticLogRecordInfo log_line_info{QUILL_STRINGIFY(__LINE__), __FILE__, __func__,
+                                    "This the {1} formatter {0}", LogLevel::Debug};
+
+  for (size_t i = 0; i < 5; ++i)
+  {
+    // Format to a buffer
+    custom_pattern_formatter.format(ts, thread_id, logger_name.data(), log_line_info, "pattern", 1234);
+    auto const& formatted_buffer = custom_pattern_formatter.formatted_log_record();
+
+    // Convert the buffer to a string
+    std::string const formatted_string = fmt::to_string(formatted_buffer);
+
+    std::string const expected_string =
+      "2020-01-23T21:42:41.099220+0000GMT [31341] PatternFormatterTest.cpp:193 LOG_DEBUG    test_logger - This the "
+      "1234 formatter pattern [TestBody]\n";
+
+    EXPECT_EQ(formatted_buffer.size(), expected_string.length());
+    EXPECT_EQ(formatted_string, expected_string);
+  }
+}
+
+TEST(PatternFormatter, custom_pattern_timestamp_strftime_reallocation_when_adding_fractional_seconds)
+{
+  // set a timestamp_format that will cause timestamp _formatted_date to re-allocate.
+  PatternFormatter custom_pattern_formatter{
+    QUILL_STRING(
+      "%(ascii_time) [%(thread)] %(filename):%(lineno) LOG_%(level_name) %(logger_name) - "
+      "%(message) [%(function_name)]"),
+    "%FT%T.%T.%Qus%z%Z", Timezone::GmtTime};
+
+  std::chrono::nanoseconds ts{1579815761099220000};
+  char const* thread_id = "31341";
+  std::string const logger_name = "test_logger";
+  StaticLogRecordInfo log_line_info{QUILL_STRINGIFY(__LINE__), __FILE__, __func__,
+                                    "This the {1} formatter {0}", LogLevel::Debug};
+
+  for (size_t i = 0; i < 5; ++i)
+  {
+    // Format to a buffer
+    custom_pattern_formatter.format(ts, thread_id, logger_name.data(), log_line_info, "pattern", 1234);
+    auto const& formatted_buffer = custom_pattern_formatter.formatted_log_record();
+
+    // Convert the buffer to a string
+    std::string const formatted_string = fmt::to_string(formatted_buffer);
+
+    std::string const expected_string =
+      "2020-01-23T21:42:41.21:42:41.099220+0000GMT [31341] PatternFormatterTest.cpp:226 LOG_DEBUG    test_logger - This the "
+      "1234 formatter pattern [TestBody]\n";
+
+    EXPECT_EQ(formatted_buffer.size(), expected_string.length());
+    EXPECT_EQ(formatted_string, expected_string);
+  }
+}
+
 TEST(PatternFormatter, invalid_pattern)
 {
 #if defined(QUILL_NO_EXCEPTIONS)